# packages/stock-data/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GlobalStockDataService",
    "description": "面向股票分析、量化交易与策略回测的全球股票数据服务 (50GB空间极致优化架构)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/stock/profile": {
      "get": {
        "tags": [
          "Advanced Analytics: Sector, Holders, LHB & Macro"
        ],
        "summary": "Get Company Profile",
        "description": "获取公司画像与所属行业分类、主营业务及上市信息：\n- A 股：对接巨潮资讯官方上市公司画像与行业归类。\n- 美股/港股：对接 Yahoo Finance 官方行业分类 (Sector / Industry)。",
        "operationId": "get_company_profile_api_v1_stock_profile_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "股票代码，如 002594 (比亚迪), 600519 (茅台), AAPL (苹果)",
              "title": "Symbol"
            },
            "description": "股票代码，如 002594 (比亚迪), 600519 (茅台), AAPL (苹果)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stock/shareholders": {
      "get": {
        "tags": [
          "Advanced Analytics: Sector, Holders, LHB & Macro"
        ],
        "summary": "Get Shareholders",
        "description": "获取股东户数（筹码集中度）与十大流通股东名单：\n- 股东总数与户均持股量（用于判断散户交出筹码、庄家机构建仓）\n- 十大股东/流通股东名单、持股数与持股占比",
        "operationId": "get_shareholders_api_v1_stock_shareholders_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "股票代码，如 002594 (比亚迪), 600519 (茅台)",
              "title": "Symbol"
            },
            "description": "股票代码，如 002594 (比亚迪), 600519 (茅台)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareholdersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/sectors": {
      "get": {
        "tags": [
          "Advanced Analytics: Sector, Holders, LHB & Macro"
        ],
        "summary": "Get Market Sectors",
        "description": "获取今日全市场板块行情热点排行：\n- 行业板块排行（申万行业分类热度）\n- 概念板块排行（华为汽车、光伏、低空经济、AI芯片等题材热点）\n- 各板块对应涨跌幅与领涨龙头股表现",
        "operationId": "get_market_sectors_api_v1_market_sectors_get",
        "parameters": [
          {
            "name": "indicator",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(行业|概念)$",
              "description": "板块类型: '行业' 或 '概念'",
              "default": "行业",
              "title": "Indicator"
            },
            "description": "板块类型: '行业' 或 '概念'"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "返回数量",
              "default": 30,
              "title": "Limit"
            },
            "description": "返回数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectorRankResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/dragon-tiger": {
      "get": {
        "tags": [
          "Advanced Analytics: Sector, Holders, LHB & Macro"
        ],
        "summary": "Get Dragon Tiger List",
        "description": "获取每日龙虎榜数据：\n机构专用席位买卖、游资营业部大举建仓或抛售的个股及上榜原因（涨跌偏离值达7%、换手率达20%等）。\n数据源: 东方财富 (stock_lhb_detail_em)，稳定且含\"上榜原因\"字段。",
        "operationId": "get_dragon_tiger_list_api_v1_market_dragon_tiger_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "交易日期 (YYYYMMDD 或 YYYY-MM-DD)，不传默认最近一个交易日",
              "title": "Date"
            },
            "description": "交易日期 (YYYYMMDD 或 YYYY-MM-DD)，不传默认最近一个交易日"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DragonTigerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/macro/china/pmi": {
      "get": {
        "tags": [
          "Advanced Analytics: Sector, Holders, LHB & Macro"
        ],
        "summary": "Get China Pmi",
        "description": "获取中国官方制造业/非制造业 PMI 采购经理人指数月度序列 (宏观景气度核心先行指标)。\n数据源: 国家统计局 (AkShare macro_china_pmi)。",
        "operationId": "get_china_pmi_api_v1_macro_china_pmi_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 300,
              "minimum": 1,
              "description": "返回最近 N 个月数据 (默认24个月)",
              "default": 24,
              "title": "Limit"
            },
            "description": "返回最近 N 个月数据 (默认24个月)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PMIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/macro/china/cpi": {
      "get": {
        "tags": [
          "Advanced Analytics: Sector, Holders, LHB & Macro"
        ],
        "summary": "Get China Cpi",
        "description": "获取中国居民消费价格指数 (CPI) 月度序列，含全国/城市/农村同比与环比增速。\n数据源: 国家统计局 (AkShare macro_china_cpi)。",
        "operationId": "get_china_cpi_api_v1_macro_china_cpi_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 300,
              "minimum": 1,
              "description": "返回最近 N 个月数据 (默认24个月)",
              "default": 24,
              "title": "Limit"
            },
            "description": "返回最近 N 个月数据 (默认24个月)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CPIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/macro/china/ppi": {
      "get": {
        "tags": [
          "Advanced Analytics: Sector, Holders, LHB & Macro"
        ],
        "summary": "Get China Ppi",
        "description": "获取中国工业生产者出厂价格指数 (PPI) 月度序列 (企业端通胀/通缩核心先行指标)。\n数据源: 国家统计局 (AkShare macro_china_ppi)。",
        "operationId": "get_china_ppi_api_v1_macro_china_ppi_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 300,
              "minimum": 1,
              "description": "返回最近 N 个月数据 (默认24个月)",
              "default": 24,
              "title": "Limit"
            },
            "description": "返回最近 N 个月数据 (默认24个月)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PPIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/macro/china/m2": {
      "get": {
        "tags": [
          "Advanced Analytics: Sector, Holders, LHB & Macro"
        ],
        "summary": "Get China M2",
        "description": "获取中国 M2 货币供应量同比增速序列 (流动性宽紧核心宏观指标)。\n数据源: 国家统计局 (AkShare macro_china_m2_yearly)。",
        "operationId": "get_china_m2_api_v1_macro_china_m2_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 300,
              "minimum": 1,
              "description": "返回最近 N 期数据 (默认24期)",
              "default": 24,
              "title": "Limit"
            },
            "description": "返回最近 N 期数据 (默认24期)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/M2Response"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyProfileResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry",
            "description": "所属行业 (如：汽车制造业、半导体等)"
          },
          "main_business": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Business",
            "description": "主营业务范围"
          },
          "listing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Date",
            "description": "上市日期"
          },
          "registered_capital": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registered Capital",
            "description": "注册资本"
          },
          "intro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intro",
            "description": "公司机构简介"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "ticker",
          "company_name"
        ],
        "title": "CompanyProfileResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SectorRankResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SectorRankItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "type",
          "count",
          "data"
        ],
        "title": "SectorRankResponse"
      },
...
      "ShareholdersResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "description": "统计报告期"
          },
          "publish_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Date",
            "description": "信息披露日期"
          },
          "total_shareholders": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Shareholders",
            "description": "股东总户数"
          },
          "avg_shares_per_holder": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Shares Per Holder",
            "description": "户均持股数量 (筹码集中度关键指标)"
          },
          "top_holders": {
            "items": {
              "$ref": "#/components/schemas/ShareholderItem"
            },
            "type": "array",
            "title": "Top Holders"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "end_date",
          "top_holders"
        ],
        "title": "ShareholdersResponse"
      },
...
    }
  }
}